Move eager flow test to pure core
